Greeting Controller add

+++ skeleton-web-services/src/main/java/org/example/ws/modal/Greeting.java
@@ -1,21 +1,19 @@
 package org.example.ws.modal;
 
-import java.math.BigInteger;
-
 public class Greeting {
 
-    private BigInteger id;
+    private Long id;
     private String text;
 
     public Greeting() {
 
     }
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
